+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `period` will be documented in this file
 
+## Unreleased
+
+- Add `Period::fromDatePeriod(DatePeriod $period): Period`
+- Add `asDatePeriod(): DatePeriod` to `Period`
+
 ## 1.4.4 - 2019-08-05
 
 - ~Performance improvement in `Period::contains()` (#46)~ edit: this change wasn't merged and targeted at 2.0